pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "NLP_Extractor"
version = "0.1.0"
description = "Description of your project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
openpyxl = "^3.1.2"
python-pptx = "^0.6.23"
python-docx = "^1.1.0"
ipykernel = "^6.28.0"
ginza = "^5.1.3"
ja-ginza = "^5.1.3"
